move packages and time slots into a Task component
