+++ js/app/components/delivery-form/Task.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useMemo, useState } from 'react'
 import { Field } from 'formik'
 import AddressBookNew from './AddressBook'
 import { Input, Button } from 'antd'
@@ -10,6 +10,10 @@ import './Task.scss'
 import TagsSelect from '../TagsSelect'
 import { TaskDateTime } from './TaskDateTime'
 import { useDeliveryFormFormikContext } from './hooks/useDeliveryFormFormikContext'
+import {
+  useGetStorePackagesQuery,
+  useGetStoreTimeSlotsQuery,
+} from '../../api/slice'
 
 
 export default ({
@@ -20,9 +24,7 @@ export default ({
   storeDeliveryInfos,
   onRemove,
   showRemoveButton,
-  packages,
   tags,
-  timeSlotLabels,
 }) => {
   const { t } = useTranslation()
 
@@ -38,6 +40,23 @@ export default ({
   const [showLess, setShowLess] = useState(
     taskValues.type === 'DROPOFF' && values.tasks.length > 2,
   )
+
+  const { data: timeSlotsData } = useGetStoreTimeSlotsQuery(storeNodeId)
+  const { data: packagesData } = useGetStorePackagesQuery(storeNodeId)
+
+  const timeSlotLabels = useMemo(() => {
+    if (timeSlotsData) {
+      return timeSlotsData['hydra:member']
+    }
+    return []
+  }, [timeSlotsData])
+
+  const packages = useMemo(() => {
+    if (packagesData) {
+      return packagesData['hydra:member']
+    }
+    return null
+  }, [packagesData])
 
   return (
     <div className="task border p-4 mb-4" data-testid={`form-task-${index}`}>
